Replace loading component with slot
