+++ app/components/primer/select_menu_component.html.erb
@@ -28,6 +28,12 @@
         <% end %>
       <% end %>
 
+      <% if loading %>
+        <%= render loading.component do %>
+          <%= loading.content %>
+        <% end %>
+      <% end %>
+
       <%= modal.content %>
     <% end %>
