+++ test/components/component_test.rb
@@ -35,7 +35,6 @@ class PrimerComponentTest < Minitest::Test
     end],
     [Primer::SelectMenuBlankslateComponent, {}],
     [Primer::SelectMenuItemComponent, {}],
-    [Primer::SelectMenuLoadingComponent, {}],
     [Primer::SpinnerComponent, {}],
     [Primer::StateComponent, { title: "Open" }],
     [Primer::SubheadComponent, { heading: "Foo" }, proc { |component| component.slot(:heading) { "Foo" } }],
